requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
python-dotenv
flask-migrate
faker
